CombinedLibraryManager to handle adding/deleting items in datagrid
I've concluded that multibindings for the datagrid itemssource just wasn't going to work. It would lose its sorting when adding or removing items and there is no way to fix that with a multibinding.

So this class acts as a layer between so the datagrid can bind to one item in the datacontext -- the CombinedLibraryManager instance.

This class is responsible for listening for change events and modifying itself when appropriate.

Bindings and converters have been changed as required.

This actually ended up requiring much less code elsewhere and should be more maintainable.